requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-telegram-bot[job-queue]==21.6
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
alembic==1.13.3
pydantic-settings==2.5.2
pydantic==2.9.2
pyyaml==6.0.2
python-dotenv==1.0.1
